+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 
 ## [UNRELEASED]
 
+### Added
+
+- Added a command `/patterns lehmer` to generate and display a Lehmer code for rearranging the stack using Swindler's Gambit.
+
 ### Mods Updated
 
 - Hexchanting: 1.1.7